% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{anomaly_smoothing}
\alias{anomaly_smoothing}
\title{Check if time series has anomalies and smoothes them.}
\usage{
anomaly_smoothing(timeseries)
}
\arguments{
\item{timeseries}{A ts object.}
}
\value{
A dataframe
}
\description{
Check if time series has anomalies and smoothes them.
}
\details{
This function allows the user to check if a time series has anomalies (outliers), both visually
also visualizing them. Then a rolling mean smoothing filter is applied to deal
with them (smoothed). The result is a smoothed time series. We also
visualize the levels, ACF and PACF plots of the comparison before and after
smoothing
}
\examples{
df <- data.frame(
        time = c(1996:2016),
        CLVMNACSCAB1GQEL_PC1 = runif(n = 21, min = 1, max = 10)
)
ts <- stats::ts(df[,"CLVMNACSCAB1GQEL_PC1"],start=c(1996,01),frequency=4)
anomaly_smoothing(ts)
}
