% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{rGDP for Greece data}
\format{
## `who`
A time series object with 113 rows and 1 column. The time stamp is encoded for
each row:
\describe{
  \item{timeseries}{Value for each quarter}
}
}
\source{
<https://is.gd/e6wKRk>
}
\usage{
timeseries
}
\description{
A simple time series dataset of real GDP data for Greece.
}
\keyword{datasets}
