% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{train_and_forecast_arima}
\alias{train_and_forecast_arima}
\title{Selects best ARIMA model and trains it based on our time series data, then does one-step ahead forecast.}
\usage{
train_and_forecast_arima(dataframe, ts)
}
\arguments{
\item{dataframe}{A pandas dataframe containing at least one ts object.}

\item{ts}{A ts object to be selected from the dataframe}
}
\value{
A dataframe
}
\description{
Selects best ARIMA model and trains it based on our time series data, then does one-step ahead forecast.
}
\details{
This function allows the user to fit an ARIMA (Auto Regression Integral Moving Average)
to a time series (training subset) using an automated function. This model is then used to do one step ahead
forecasts of the available time series (test subset). The forecast results are saved and displayed
compared to the original.
}
\examples{
df <- data.frame(
        time = c(1996:2016),
        CLVMNACSCAB1GQEL_PC1 = runif(n = 21, min = 1, max = 10)
)
name <- "CLVMNACSCAB1GQEL_PC1"
train_and_forecast_arima(df, name)
}
